check to if nifi is downloaded and don't redownload

+++ install_nifi.sh
@@ -6,7 +6,11 @@ NIFI_DIR=/opt/nifi
 CHECKSUM_PASS=good
 MIN_JAVA="-Xms1024m"
 MAX_JAVA="-Xmx1024m"
-wget http://apache.mirror.anlx.net/nifi/$GET_VERSION -O /tmp/${NIFI_TAR}
+if [[ ! -e "/tmp/${GET_VERSION}" ]]
+    then
+        wget http://apache.mirror.anlx.net/nifi/$GET_VERSION -O /tmp/${NIFI_TAR}
+    fi
+sudo apt -y install openjdk-8-jre-headless
 # the below returns the checksum and the file name then converts to bash array
 checksum=$(sha256sum "/tmp/${NIFI_TAR}"| cut -d' ' -f1)
 checksum_test=$(curl -L "https://checker.apache.org/mk_page.cgi?CSUM=${checksum}&BUT_CSUM=search" | grep  "the file has a good.*signature" | grep -o "good" )
@@ -22,3 +26,4 @@ sudo chown -R ${sudo_user}:${sudo_user} ${NIFI_DIR}
 tar xzf /tmp/${NIFI_TAR} --strip-components=1 -C ${NIFI_DIR}
 cp ${NIFI_DIR}/conf/bootstrap.conf ${NIFI_DIR}/conf/bootstrap.orig
 sed -i -e "/java.arg.2=/ s/=.*/=${MIN_JAVA}/" -e "/java.arg.3=/ s/=.*/=${MAX_JAVA}/" ${NIFI_DIR}/conf/bootstrap.conf
+/opt/nifi/bin/nifi.sh start
